test(storage): scope windows stronghold helpers
